% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{align.optim}
\alias{align.optim}
\title{Runs alignment optimizer}
\usage{
align.optim(
  stacked,
  n,
  estimator,
  nstarts = 50,
  ncores = 3,
  hyper.first,
  center.means,
  eps.alignment,
  clf.ignore.quantile
)
}
\description{
Not generally intended to be used on its own, but exported anyway
for didactic purposes.
}
\details{
See example for \code{\link{Alignment}} for examples
}
