% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{loadEstimates.lavaan.ordered}
\alias{loadEstimates.lavaan.ordered}
\title{Estimate \code{lavaan} models using aligned parameter estimates}
\usage{
loadEstimates.lavaan.ordered(
  fit,
  align.mean,
  align.variance,
  newpars,
  do.fit = T
)
}
\description{
Not generally intended to be used on its own, but exported anyway for
didactic purposes.
}
\details{
See example for \code{\link{Alignment}} for examples

This may differ from what is used in Mplus.
}
