% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stackEstimates}
\alias{stackEstimates}
\title{Stack estimates for optimization}
\usage{
stackEstimates(estList)
}
\arguments{
\item{estList}{List of estimates from \code{\link{getEstimates.lavaan}} or
\code{\link{getEstimates.mirt}} to stack to feed into
\code{\link{SF.mplus3D}}}
}
\value{
A set of estimates prepared for efficient use with
\code{\link{SF.mplus3D}}
}
\description{
Not generally intended to be used on its own, but exported anyway for
didactic purposes.
}
\details{
See example for \code{\link{Alignment}} for examples
}
