% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{transformEstimates.lavaan.ordered}
\alias{transformEstimates.lavaan.ordered}
\title{Transform \code{lavaan} estimates using aligned estimates of latent
mean and variance}
\usage{
transformEstimates.lavaan.ordered(
  align.mean,
  align.variance,
  est,
  toCompare = FALSE
)
}
\arguments{
\item{align.mean}{Mean to transform model to.}

\item{align.variance}{Variance to transform model to.}

\item{est}{Estimates to transform, from \code{\link{getEstimates.lavaan}}}

\item{toCompare}{Accounts for discrepancies between delta and theta
parameterizations; see \code{\link{Alignment}} documentation.}
}
\value{
Estimates in the same structure as from
\code{\link{getEstimates.lavaan}}, but transformed from (assumed) mean 0 and
variance 1 to the metric specified by \code{align.mean} and \code{align.variance}.
}
\description{
Not generally intended to be used on its own, but exported anyway for
didactic purposes.
}
\details{
See example for \code{\link{Alignment}} for examples

This program was designed based on the published work of Asparouhov & Muthen,
and was not intended to match Mplus exactly, and may not.
}
