% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getEstimates.lavaan}
\alias{getEstimates.lavaan}
\title{Prepare \code{lavaan} estimates for alignment}
\usage{
getEstimates.lavaan(fit, SE = TRUE)
}
\description{
Not generally intended to be used on its own, but exported anyway for didactic purposes.
}
\details{
See example for \code{\link{Alignment}} for examples

This may differ from what is used in Mplus.
}
