% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getEstimates.mirt}
\alias{getEstimates.mirt}
\title{Prepare \code{mirt} estimates for alignment}
\usage{
getEstimates.mirt(fit, SE = FALSE, bifactor.marginal = F)
}
\description{
Not generally intended to be used on its own, but exported anyway for didactic purposes.
}
\details{
See example for \code{\link{Alignment}} for examples

This may differ from what is used in Mplus.
}
