% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{loadEstimates.mirt.grm}
\alias{loadEstimates.mirt.grm}
\title{Estimate \code{mirt} models using aligned parameter estimates}
\usage{
loadEstimates.mirt.grm(fit, align.mean, align.variance, newpars, do.fit = T)
}
\description{
Not generally intended to be used on its own, but exported anyway for didactic purposes.
}
\details{
See example for \code{\link{Alignment}} for examples

This may differ from what is used in Mplus.
}
