% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{SF.mplus3D}
\alias{SF.mplus3D}
\title{Simplicity function for alignment}
\usage{
SF.mplus3D(
  pars,
  est,
  comb,
  nobs,
  estimator,
  eps.alignment,
  clf.ignore.quantile,
  hyper = "all",
  otherHyper = NULL
)
}
\arguments{
\item{pars}{Hyperparameters to feed into optimizer}

\item{est}{Estimates to transform, from \code{\link{getEstimates.mirt}} or
\code{\link{getEstimates.lavaan}}}

\item{comb}{All combinations of groups from \code{\link[utils]{combn}}}

\item{nobs}{Sample size in each group}

\item{estimator}{See \code{\link{Alignment}} documentation.}

\item{eps.alignment}{See \code{\link{Alignment}} documentation.}

\item{clf.ignore.quantile}{See \code{\link{Alignment}} documentation.}

\item{hyper}{Hyperparameter to calculate simplicity function for; see
\code{\link{Alignment}} documentation.}

\item{otherHyper}{Non-included hyperparameter}
}
\value{
A value of the simplicity function from Asparouhuv & Muthen, 2014.
}
\description{
Not generally intended to be used on its own, but exported anyway for
didactic purposes.
}
\details{
See example for \code{\link{Alignment}} for examples

This program was designed based on the published work of Asparouhov & Muthen,
and was not intended to match Mplus exactly, and may not.
}
