% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stackEstimates}
\alias{stackEstimates}
\title{Stack estimates for optimization}
\usage{
stackEstimates(estList)
}
\description{
Not generally intended to be used on its own, but exported anyway for didactic purposes.
}
\details{
See example for \code{\link{Alignment}} for examples

This may differ from what is used in Mplus.
}
