% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{transformEstimates.lavaan.ordered}
\alias{transformEstimates.lavaan.ordered}
\title{Transform \code{lavaan} estimates using aligned estimates of latent
mean and variance}
\usage{
transformEstimates.lavaan.ordered(
  align.mean,
  align.variance,
  est,
  toCompare = F
)
}
\arguments{
\item{align.mean}{Mean to transform model to.}

\item{align.variance}{Variance to transform model to.}

\item{est}{Estimates to transform, from \code{\link{getEstimates.lavaan}}}

\item{toCompare}{Accounts for discrepancies between delta and theta
parameterizations; see \code{\link{Alignment}} documentation.}
}
\description{
Not generally intended to be used on its own, but exported anyway for
didactic purposes.
}
\details{
See example for \code{\link{Alignment}} for examples

This may differ from what is used in Mplus.
}
