% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{transformEstimates.mirt.grm}
\alias{transformEstimates.mirt.grm}
\title{Transform \code{mirt} estimates using aligned estimates of latent mean
and variance}
\usage{
transformEstimates.mirt.grm(align.mean, align.variance, est)
}
\description{
Not generally intended to be used on its own, but exported anyway for
didactic purposes.
}
\details{
See example for \code{\link{Alignment}} for examples

This may differ from what is used in Mplus.
}
